returns the number to be fliped to get from one to another

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int result, ret;
+	unsigned int long result, ret;
 
 	result = n ^ m;
